requirements.txt
connexion[swagger-ui]
Flask
flask-sqlalchemy
python-dateutil
requests